+++ python/ray/tune/schedulers/pbt.py
@@ -12,7 +12,7 @@ from ray.tune.error import TuneError
 from ray.tune.result import DEFAULT_METRIC, TRAINING_ITERATION
 from ray.tune.search import SearchGenerator
 from ray.tune.utils.util import SafeFallbackEncoder
-from ray.tune.sample import Domain, Function
+from ray.tune.search.sample import Domain, Function
 from ray.tune.schedulers import FIFOScheduler, TrialScheduler
 from ray.tune.search.variant_generator import format_vars
 from ray.tune.experiment import Trial
