+++ python/ray/tune/search/hebo/hebo_search.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 
 from ray.tune.result import DEFAULT_METRIC
-from ray.tune.sample import (
+from ray.tune.search.sample import (
     Categorical,
     Domain,
     Float,
