+++ python/ray/tune/search/optuna/optuna_search.py
@@ -7,7 +7,7 @@ from packaging import version
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 from ray.tune.result import DEFAULT_METRIC, TRAINING_ITERATION
-from ray.tune.sample import (
+from ray.tune.search.sample import (
     Categorical,
     Domain,
     Float,
