+++ python/ray/tune/search/skopt/skopt_search.py
@@ -5,7 +5,14 @@ import pickle
 from typing import Dict, List, Optional, Tuple, Union, Any
 
 from ray.tune.result import DEFAULT_METRIC
-from ray.tune.sample import Categorical, Domain, Float, Integer, Quantized, LogUniform
+from ray.tune.search.sample import (
+    Categorical,
+    Domain,
+    Float,
+    Integer,
+    Quantized,
+    LogUniform,
+)
 from ray.tune.search import Searcher
 from ray.tune.search import (
     UNRESOLVED_SEARCH_SPACE,
