+++ python/ray/tune/search/variant_generator.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, Generator, Iterable, List, Optional, Tuple
 import numpy
 import random
 
-from ray.tune.sample import Categorical, Domain, Function, RandomState
+from ray.tune.search.sample import Categorical, Domain, Function, RandomState
 from ray.util.annotations import DeveloperAPI
 
 logger = logging.getLogger(__name__)
