+++ python/ray/tune/suggest/suggestion.py
@@ -1,9 +1,9 @@
 from ray.tune._structure_refactor import warn_structure_refactor
-from ray.tune.search import (
+from ray.tune.search import (  # noqa: F401, F403
     UNDEFINED_SEARCH_SPACE,
     UNRESOLVED_SEARCH_SPACE,
     UNDEFINED_METRIC_MODE,
-)  # noqa: F401, F403
+)
 from ray.tune.search.searcher import Searcher  # noqa: F401, F403
 from ray.tune.search.concurrency_limiter import ConcurrencyLimiter  # noqa: F401, F403
